Lock other map in shared mode for copy assignment
In the copy assignment operator, we do not need to take an exclusive
lock on the map we are copying from since our reference to it is const.
Passing an *unlocked* std::shared_lock as a wrapper around the other's
mutex lets std::scoped_lock lock it in shared mode.